refactor: move tflite::Span to independent header
Hoist the universally useful tflite::Span out of
codegen/runtime/micro_codegen_context.h and into an independent header
usable from elsewhere in the project.
